Разбил 1 файл на 3, добавав модульность. main. js - точка входа - модуль, который связывает другие модули, util.js - модуль с вспомогательными функциями, data.js - модуль, который создаёт данные.
