+++ js/data.js
@@ -1,0 +1,73 @@
+import{getRandomInteger,getRandomArrayElement,getRandomUniqeValueFromRage,createComments} from './util.js';
+
+//массив имен для создания объектов
+
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+  'Леонардо',
+  'Донаттело',
+  'Рафаэль',
+  'Микеланджело',
+  'Антон',
+  'Евгений',
+  'Даниил',
+  'Илья'
+];
+
+//массив описаний для фотографии
+
+const DESCRIPTIONS = [
+  'Красивая фотография',
+  'Некрасивая фотография',
+  'Сгенерированная нейросетью фотография'
+];
+
+//массив комментарий
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const generatePhotoId = getRandomUniqeValueFromRage(1, 25);
+const generateUrl = getRandomUniqeValueFromRage(1, 25);
+const commentId = getRandomUniqeValueFromRage(1, 250);
+
+/**
+ * Создание объекта фотографии
+ * @returns объект фотографии
+ */
+function createPhotoObject () {
+  return {
+    id: generatePhotoId(), //идентификатор опубликованной фотографии. От 1 до 25. Id не должны повторяться
+    url: `photos/${generateUrl()}.jpg`, //адрес картинки вида photos/{{i}}.jpg, где {{i}} — число от 1 до 25. Адреса картинок не должны повторяться
+    description: getRandomArrayElement(DESCRIPTIONS), //описание фотографии. Описание придумайте самостоятельно
+    likes: getRandomInteger(15,200), //количество лайков, поставленных фотографии. Случайное число от 15 до 200
+    comments: [{ //список комментариев, оставленных другими пользователями к этой фотографии. Количество комментариев к каждой фотографии — случайное число от 0 до 30. Все комментарии генерируются случайным образом
+      id: commentId(), //у каждого комментария есть идентификатор — id — любое число. Идентификаторы не должны повторяться
+      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`, //строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg. Аватарки подготовлены в директории img
+      message: createComments(COMMENT_MESSAGES), //Для формирования текста комментария — message — необходимо взять одно или два случайных предложения из перечня
+      name: getRandomArrayElement(NAMES),//имена авторов должны быть случайными
+
+    }]
+  };
+}
+
+//Количество объектов массива — описание фотографий, опубликованной пользователем
+const PHOTO_COUNT = 25;
+
+//Генерация массива фотографий
+const createPhotoObjects = () => Array.from({length: PHOTO_COUNT}, createPhotoObject);
+
+export {createPhotoObjects};
